[ODPM-212] Factor out common `get_years` and `get_totals` functions
